requirements.txt
flask 
peewee
python-dotenv
psycopg2-binary